refactor: 게시글 <- 카테고리 패키지 통합
