+++ src/main/java/com/bob/domain/post/service/reader/CategoryReader.java
@@ -1,7 +1,7 @@
-package com.bob.domain.category.service.reader;
+package com.bob.domain.post.service.reader;
 
-import com.bob.domain.category.entity.Category;
-import com.bob.domain.category.repository.CategoryRepository;
+import com.bob.domain.post.entity.Category;
+import com.bob.domain.post.repository.CategoryRepository;
 import com.bob.global.exception.exceptions.ApplicationException;
 import com.bob.global.exception.response.ApplicationError;
 import java.util.List;
